--- src/main/java/com/kusitms/wannafly/command/auth/infrastructure/refreshtoken/JpaRefreshTokenRepository.java
@@ -1,7 +1,0 @@
-package com.kusitms.wannafly.command.auth.infrastructure.refreshtoken;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaRefreshTokenRepository extends JpaRepository<JpaRefreshToken, String> {
-
-}
