+++ src/main/java/com/kusitms/wannafly/command/auth/infrastructure/refreshtoken/RefreshTokenRedisRepository.java
@@ -1,0 +1,6 @@
+package com.kusitms.wannafly.command.auth.infrastructure.refreshtoken;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface RefreshTokenRedisRepository extends CrudRepository<RedisRefreshToken, String> {
+}
