+++ src/main/java/com/kusitms/wannafly/command/auth/infrastructure/refreshtoken/RefreshTokenRepositoryImpl.java
@@ -4,33 +4,29 @@ import com.kusitms.wannafly.command.auth.token.RefreshToken;
 import com.kusitms.wannafly.command.auth.token.RefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
 public class RefreshTokenRepositoryImpl implements RefreshTokenRepository {
 
-    private final JpaRefreshTokenRepository jpaRefreshTokenRepository;
+    private final RefreshTokenRedisRepository refreshTokenRedisRepository;
+
 
     @Override
-    @Transactional
     public void save(RefreshToken refreshToken) {
-        JpaRefreshToken jpaRefreshToken = new JpaRefreshToken(refreshToken);
-        jpaRefreshTokenRepository.save(jpaRefreshToken);
+        refreshTokenRedisRepository.save(new RedisRefreshToken(refreshToken));
     }
 
     @Override
     public Optional<RefreshToken> findByValue(String refreshTokenValue) {
-        return jpaRefreshTokenRepository.findById(refreshTokenValue)
-                .map(JpaRefreshToken::toRefreshToken);
+        return refreshTokenRedisRepository.findById(refreshTokenValue)
+                .map(RedisRefreshToken::toRefreshToken);
     }
 
     @Override
-    @Transactional
     public void delete(RefreshToken refreshToken) {
-        jpaRefreshTokenRepository.delete(new JpaRefreshToken(refreshToken));
+        refreshTokenRedisRepository.delete(new RedisRefreshToken(refreshToken));
     }
 }
