+++ src/main/java/com/kusitms/wannafly/command/auth/token/RefreshTokenProvider.java
@@ -4,13 +4,11 @@ import com.kusitms.wannafly.exception.BusinessException;
 import com.kusitms.wannafly.exception.ErrorCode;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Component
-@Transactional(readOnly = true)
 public class RefreshTokenProvider {
 
     private static final String EXPIRE_LENGTH = "${security.refresh.token.expire-day}";
@@ -24,7 +22,6 @@ public class RefreshTokenProvider {
         this.refreshTokenRepository = refreshTokenRepository;
     }
 
-    @Transactional
     public RefreshToken createToken(TokenPayload payload) {
         String value = UUID.randomUUID().toString();
         LocalDateTime now = LocalDateTime.now();
@@ -33,12 +30,10 @@ public class RefreshTokenProvider {
         return refreshToken;
     }
 
-    @Transactional
     public RefreshToken reIssueToken(String refreshTokenValue) {
         RefreshToken refreshToken = getRefreshToken(refreshTokenValue);
         RefreshToken newRefreshToken = createToken(new TokenPayload(refreshToken.getMemberId()));
         refreshTokenRepository.delete(refreshToken);
-        refreshTokenRepository.save(newRefreshToken);
         return newRefreshToken;
     }
 
@@ -51,7 +46,6 @@ public class RefreshTokenProvider {
         return refreshToken;
     }
 
-    @Transactional
     public void deleteToken(String refreshTokenValue) {
         refreshTokenRepository.findByValue(refreshTokenValue)
                 .ifPresent(refreshTokenRepository::delete);
