+++ src/main/resources/application-local.yml
@@ -10,6 +10,9 @@ spring:
       hibernate:
         format_sql: true
         show_sql: true
+  redis:
+    host: localhost
+    port: 6379
   security:
     oauth2:
       client:
