+++ src/main/resources/application.yml
@@ -9,9 +9,6 @@ spring:
     enabled: true
     baseline-on-migrate: true
     version: 1
-  redis:
-    host: redis
-    port: 6379
 
 
 security:
